Fix test suite failing on ruby 1.8 after 175523d6879353596310bedaefb06d12fe35049c
After the above commit, the generated Shared Strings Table (SST) no
longer followed the order the strings appeared in the
worksheets/rows/cells.  This was a side-effect of switching the inner
structure from an array to a hash - hashes in ruby 1.8 don't preserve
insertion order.

This is in and of itself not a bug nor a problem.  The generated XLS are
perfectly legal and open fine in Excel and other parsers.  There's no
technical requirement, from my readings, that says the string at
sheet[0] row[0] col[0] must be inserted at SST[0]

However, the test suite makes that assumption and there are many
instances on relying on the ordering.

This commit takes the approach of, instead of large changes to the test
suite, to introduce order back into the SST generation.  This is done
with minor changes to the hash structure, without needing to revert back
to an array and previous slow performance the hash structure fixed.